RequestEngine/gcp_cloudrun/ane1/funcfiles/requirements.txt
google-cloud-secret-manager==2.*
requests==2.*
gunicorn==21.*
flask==3.*
